laptop_inventory/requirements.txt
flask
flask-mysqldb
pymongo
flask-bcrypt
python-dotenv